+++ src/models/thruster.py
@@ -9,7 +9,7 @@ from .base import CamelCaseModel
 class ThrusterData(CamelCaseModel):
     """Model for thruster data."""
 
-    direction_vector: NDArray[Shape["8"], np.float64] | None = np.zeros(8)  # pyright: ignore[reportGeneralTypeIssues]
+    direction_vector: NDArray[Shape["8"], np.float32] | None = np.zeros(8)  # pyright: ignore[reportGeneralTypeIssues]
     last_direction_time: float = 0.0
     test_thruster: int | None = None
     test_start_time: float = 0.0
